[INJIMOB-2633]: Running existing tests
