+++ kotlin/PixelPass/build.gradle.kts
@@ -47,6 +47,7 @@ kotlin {
         }
         val commonTest by getting {
             dependencies {
+                implementation(kotlin("test"))
                 implementation(libs.junit)
                 implementation(libs.mockk)
                 implementation(libs.json)
