+++ kotlin/PixelPass/publish-artifact.gradle
@@ -3,14 +3,7 @@ version = "0.6.0-SNAPSHOT"
 
 publishing {
     repositories {
-        maven {
-            url = System.getenv("OSSRH_URL")
-            name = "pixelpass"
-            credentials {
-                username = System.getenv("OSSRH_USER")
-                password = System.getenv("OSSRH_SECRET")
-            }
-        }
+        mavenLocal()
     }
     publications {
         aar(MavenPublication) {
@@ -189,9 +182,4 @@ publishing {
             }
         }
     }
-}
-signing {
-    useGpgCmd()
-    sign publishing.publications.aar
-    sign publishing.publications.jarRelease
 }
