--- kotlin/PixelPass/src/androidMain/kotlin/io/mosip/pixelpass/Logger.kt
@@ -1,7 +1,0 @@
-package io.mosip.pixelpass
-
-import android.util.Log
-
-actual fun logMessage( message: String, tag: String){
-    Log.d(tag, message)
-}
