--- kotlin/PixelPass/src/commonMain/kotlin/io/mosip/pixelpass/Logger.kt
@@ -1,3 +1,0 @@
-package io.mosip.pixelpass
-
-expect fun logMessage(message: String, tag: String = "PixelPass::")
