--- kotlin/PixelPass/src/commonTest/java/io/mosip/qr_generator/shared/ConstantsTest.kt
@@ -1,19 +1,0 @@
-package io.mosip.qr_generator.shared
-
-import io.mockk.clearAllMocks
-import org.junit.After
-import org.junit.Assert
-import org.junit.Test
-
-class ConstantsTest{
-    @After
-    fun after(){
-        clearAllMocks()
-    }
-//    @Test
-//    fun `should test all declared constants`(){
-//        Assert.assertEquals(QR_SCALE,25)
-//        Assert.assertEquals(.\QR_BORDER,3)
-//        Assert.assertEquals(Constants.QR_QUALITY,100)
-//    }
-}
