+++ kotlin/PixelPass/src/commonTest/kotlin/io/mosip/qr_generator/shared/ConstantsTest.kt
@@ -1,0 +1,11 @@
+package io.mosip.qr_generator.shared
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class ZLibTest {
+    @Test
+    fun testCompression() {
+        assertEquals(4, 2 + 2) // Replace with actual logic
+    }
+}
