--- kotlin/PixelPass/src/jvmMain/kotlin/io/mosip/pixelpass/Logger.kt
@@ -1,5 +1,0 @@
-package io.mosip.pixelpass
-
-actual fun logMessage(message: String, tag: String){
-    println(tag+message)
-}
